+++ src/pages/_app.tsx
@@ -1,7 +1,7 @@
 import '/public/css/globals.css'
 import type { AppProps } from 'next/app'
 
-import { GoogleAnalytics } from '../components'
+import { GoogleAnalytics } from '~/components'
 
 function MyApp({ Component, pageProps }: AppProps): JSX.Element {
   return (
